+++ src/commands/music/mod.rs
@@ -1,17 +1,21 @@
+mod dashboard;
 mod disconnect;
 mod jam;
 mod join;
 mod lyrics;
 mod playing;
+mod repeat;
 mod shuffle;
 mod stop;
 mod volume;
 
+pub use dashboard::*;
 pub use disconnect::*;
 pub use jam::*;
 pub use join::*;
 pub use lyrics::*;
 pub use playing::*;
+pub use repeat::*;
 pub use shuffle::*;
 pub use stop::*;
 pub use volume::*;